+++ src/components/select/CountryMultiSelect.tsx
@@ -13,8 +13,11 @@ import { Checkbox } from 'src/components/checkbox/Checkbox';
 import { type HelpTextProps } from 'src/components/help-text/HelpText';
 import { MultiSelect } from 'src/components/select/MultiSelect';
 import { type Flag, FlagIcon } from 'src/icons/flag-icon/FlagIcon';
-import { type CountryOption, type UnavailableCountry } from 'src/types/Country';
 import type { SelectOption } from 'src/types/SelectOption';
+import {
+  type CountryOption,
+  type UnavailableCountry,
+} from 'src/utils/hooks/useCountryOptions';
 import { prepRegionCountryOptions } from 'src/utils/prepRegionCountryOptions';
 
 type AdditionalProps = {
