+++ src/components/select/CountrySelect.tsx
@@ -1,7 +1,7 @@
 import { type SelectProps, Select } from 'src/components/select/Select';
 import { type Flag, FlagIcon } from 'src/icons/flag-icon/FlagIcon';
 import type { BaseProps } from 'src/types/BaseProps';
-import type { CountryOption } from 'src/types/Country';
+import type { CountryOption } from 'src/utils/hooks/useCountryOptions';
 
 type CountrySelectType<T extends string> = {
   autoFocus?: SelectProps['autoFocus'];
