+++ src/components/select/__stories__/CountryMultiSelect.stories.tsx
@@ -7,7 +7,7 @@ import {
   type CountryMultiSelectProps,
   CountryMultiSelect,
 } from 'src/components/select/CountryMultiSelect';
-import type { CountryOption } from 'src/types/Country';
+import type { CountryOption } from 'src/utils/hooks/useCountryOptions';
 import { useCountryOptions } from 'src/utils/hooks/useCountryOptions';
 
 type RandomCountryCode = 'AD' | 'AE' | 'AF' | 'AG' | 'AI' | 'AL' | 'AT';
