+++ src/components/select/__stories__/CountryPhoneSelect.stories.tsx
@@ -8,7 +8,7 @@ import {
   CountryPhoneSelect,
 } from 'src/components/select/CountryPhoneSelect';
 import { type Flag, FlagIcon } from 'src/icons/flag-icon/FlagIcon';
-import type { CountryOption } from 'src/types/Country';
+import type { CountryOption } from 'src/utils/hooks/useCountryOptions';
 import { useCountryOptions } from 'src/utils/hooks/useCountryOptions';
 
 const CountryPhoneSelectMeta: Meta = {
